third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/media-elements/user-interface/muted-expected.txt

This is a testharness.js-based test.
PASS getting audio.muted (parser-created)
PASS setting audio.muted (parser-created)
FAIL getting audio.muted with muted="" (parser-created) assert_true: expected true got false
FAIL setting audio.muted with muted="" (parser-created) assert_equals: expected true but got false
PASS getting video.muted (parser-created)
PASS setting video.muted (parser-created)
FAIL getting video.muted with muted="" (parser-created) assert_true: expected true got false
FAIL setting video.muted with muted="" (parser-created) assert_equals: expected true but got false
PASS getting video.muted with muted="" after load (parser-created)
PASS getting audio.muted (script-created)
PASS setting audio.muted (script-created)
PASS getting audio.muted with muted="" (script-created)
PASS setting audio.muted with muted="" (script-created)
PASS getting audio.muted with muted="" (innerHTML-created)
FAIL getting audio.muted with muted="" (document.write-created) assert_true: expected true got false
PASS getting video.muted (script-created)
PASS setting video.muted (script-created)
PASS getting video.muted with muted="" (script-created)
PASS setting video.muted with muted="" (script-created)
PASS getting video.muted with muted="" (innerHTML-created)
FAIL getting video.muted with muted="" (document.write-created) assert_true: expected true got false
Harness: the test ran to completion.
